backend/requirements.txt
fastapi
python-dotenv
llama-index
llama-index-embeddings-nvidia
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-jose[cryptography]
python-multipart
requests
pydantic-settings
jwt
uvicorn
apscheduler
cryptography
sqlalchemy
asyncpg
alembic
